Start splitting code in multiple files
