+++ server/string-utils.ts
@@ -1,0 +1,3 @@
+export function isEmptyOrSpaces(str: string) {
+	return (str == null || str.trim() === '');
+}
